Moved ProblemList to a new components folder
