+++ src/analysis/retail/monk/brewmaster/modules/problems/PurifyingBrew/index.tsx
@@ -5,8 +5,8 @@ import SPELLS from 'common/SPELLS';
 import { WclTable, WCLThreatBand, WCLThreatTableResponse } from 'common/WCL_TYPES';
 import { TooltipElement } from 'interface';
 import { GuideProps, PassFailBar, SubSection } from 'interface/guide';
-import type { Problem, ProblemRendererProps } from 'interface/guide/ProblemList';
-import ProblemList from 'interface/guide/ProblemList';
+import type { Problem, ProblemRendererProps } from 'interface/guide/components/ProblemList';
+import ProblemList from 'interface/guide/components/ProblemList';
 import SpellLink from 'interface/SpellLink';
 import { AddStaggerEvent, EventType, RemoveStaggerEvent } from 'parser/core/Events';
 import { Info } from 'parser/core/metric';
